+++ source/common/router/header_formatter.cc
@@ -135,14 +135,12 @@ RequestInfoHeaderFormatter::RequestInfoHeaderFormatter(absl::string_view field_n
       return RequestInfo::Utility::formatDownstreamAddressNoPort(
           *request_info.downstreamLocalAddress());
     };
-  } else if (field_name == "START_TIMESTAMP") {
+  } else if (field_name == "START_TIME_SINCE_EPOCH") {
     field_extractor_ = [](const RequestInfo::RequestInfo& request_info) {
       // Returns the start timestamp in milliseconds.
-      return fmt::format("{}", std::chrono::duration_cast<std::chrono::milliseconds>(
-                                   request_info.startTime().time_since_epoch())
-                                   .count());
+      return AccessLog::AccessLogFormatUtils::durationToString(request_info.startTime().time_since_epoch());
     };
-  } else if (field_name.find_first_of("UPSTREAM_METADATA") == 0) {
+  } else if (field_name.find("UPSTREAM_METADATA") == 0) {
     field_extractor_ =
         parseUpstreamMetadataField(field_name.substr(STATIC_STRLEN("UPSTREAM_METADATA")));
   } else {